+++ src/js/getAlphabetFromRoot.js
@@ -1,9 +1,9 @@
 //return alphabet starting with the root of the chord
 function getAlphabetFromRoot(user_chord) {
 
-	let chord_root = user_chord[0];
-	var alpha_root = [];
-	var alpha_temp = [];
+	let chord_root = user_chord[0]
+	var alpha_root = []
+	var alpha_temp = []
 
 	_alphabet.forEach(function (key) {
 		if (key === chord_root || alpha_root.length) {
@@ -14,7 +14,7 @@ function getAlphabetFromRoot(user_chord) {
 			alpha_temp.push(key)
 		}
 		//console.log(key)
-	});
-
+	})
 	return alpha_root.concat(alpha_temp)
 }
+module.exports = getAlphabetFromRoot