+++ src/js/getChord.js
@@ -43,3 +43,4 @@ function getChord(userChord, userIntervals) {
 	//return userChord[0] + " " + output.name
 	return output
 }
+module.exports = getChord