+++ src/js/getUserChord.js
@@ -1,7 +1,0 @@
-// sample function that returned a chord
-// used to test while app was only in the console
-function getUserChord() {
-	//return ["D","F#","A","C#"];
-	//return ["D","F","A","C"];
-	return ["D", "F#", "A", "C#"]
-}